requirements.txt
flask>=3.0
pyjwt>=2.8
pandas>=2.1
openpyxl>=3.1
gunicorn>=21.2
psycopg2-binary>=2.9
